+++ beacon-chain/state/state-native/setters_withdrawal.go
@@ -32,10 +32,12 @@ func (b *BeaconState) SetWithdrawalQueue(val []*enginev1.Withdrawal) error {
 
 // AppendWithdrawal adds a new withdrawal to the end of withdrawal queue.
 // This function assumes that the caller holds a lock on b.
-func (b *BeaconState) appendWithdrawal(wal *enginev1.Withdrawal) error {
+func (b *BeaconState) AppendWithdrawal(wal *enginev1.Withdrawal) error {
 	if b.version < version.Capella {
-		return errNotSupported("appendWithdrawal", b.version)
+		return errNotSupported("AppendWithdrawal", b.version)
 	}
+	b.lock.Lock()
+	defer b.lock.Unlock()
 
 	q := b.withdrawalQueue
 	if wal == nil || wal.WithdrawalIndex != uint64(len(q)) {
@@ -98,50 +100,4 @@ func (b *BeaconState) SetNextPartialWithdrawalValidatorIndex(i types.ValidatorIn
 
 	b.nextPartialWithdrawalValidatorIndex = i
 	return nil
-}
-
-// WithdrawBalance withdraws the balance from the validator and creates a
-// withdrawal receipt for the EL to process
-func (b *BeaconState) WithdrawBalance(index types.ValidatorIndex, amount uint64) error {
-	if b.version < version.Capella {
-		return errNotSupported("WithdrawBalance", b.version)
-	}
-
-	val, err := b.ValidatorAtIndexReadOnly(index)
-	if err != nil {
-		return errors.Wrapf(err, "could not get validator at index %d", index)
-	}
-
-	// Protection against withdrawing a BLS validator, this should not
-	// happen in runtime!
-	if !val.HasETH1WithdrawalCredential() {
-		return errors.New("could not withdraw balance from validator: invalid withdrawal credentials")
-	}
-
-	b.lock.Lock()
-	defer b.lock.Unlock()
-	if uint64(index) >= uint64(len(b.balances)) {
-		return errors.New("could not withdraw balance from validator: invalid index")
-	}
-	balAtIdx := b.balances[index]
-
-	if amount > balAtIdx {
-		balAtIdx = 0
-	} else {
-		balAtIdx -= amount
-	}
-
-	b.balances[index] = balAtIdx
-	b.markFieldAsDirty(nativetypes.Balances)
-	b.addDirtyIndices(nativetypes.Balances, []uint64{uint64(index)})
-
-	withdrawal := &enginev1.Withdrawal{
-		WithdrawalIndex:  b.nextWithdrawalIndex,
-		ValidatorIndex:   index,
-		ExecutionAddress: val.WithdrawalCredentials()[12:],
-		Amount:           amount,
-	}
-
-	b.nextWithdrawalIndex += 1
-	return b.appendWithdrawal(withdrawal)
 }
